Give user option to fix failed downloads:
When a download fails, the range return is start: 0 and end: -1 when trying to load the tract.

This obviously results in an error. The fix here is to prevent crash, bounce back to home, and prompt user to download resources again.

If user chooses yes, all translations for a resource are downloaded again.