Prevent crash when using european localization
NumberFormatter assumes a comma "," when parsing numbers in some locales. When set to German, this crashed because it expects a comma, not a period for decimal separator.

The fix is to create one static NumberFormatter that can be used and always have it look for periods since that's what our XML uses