Drop localizedName column in favor of runtime method to determine the localizedName
The problem with the previous approach was that the stored value only changed when the Language.code changed. The code never changes however. The result is that the localized name value would not change when the runtime Locale changes because the code remains the same.

The fix is to stop storing the localizedName as a property (Realm empty migration required.) and instead create a method to do the conversion with the Locale defined at Runtime.