Support jumping into specific (but arbitrary) resource from app delegate
Allow jumping into a specific resource point by setting language, resource and page number in the flow controller. The set language overrides the primary language just for the current view of that tract. Parallel language still works if it is set.

Add a function in TractViewController to "resolve" primary language, by taking the one set as universalLink language instead of the primary language if it is set.