Fix rendering of navigation buttons in XCode 9
The issue actually isn't specific to the "right side" buttons, though the issue was noticed there.

The reason is because the buttons that are shown on the right side of the bar have images that are signficantly larger than the left side.

This version of XCode that builds the project doesn't respect the size applied to UIBarButton item and allows the image (in customView) to expand as large as it wants to. The old builder enforced the limit at the UIBarButtonItem level

The actual fix is to apply height and width constraints to the buttons inside the UIBarButtonItem to make sure they stay 22x22 regardless of the size of the image they contain.