Remove linguistic tagger in favor of hard coded R2L languages
the linguistic tagger didn't always get it right. instead at the language level, hard code if the language is right to left. If so, when calculating the alignment, flip left/right in cases when the languages is left to right.

this implies the default for the app is left to right. note we intentionally don't use NSTextAlignment.natural because that only works on system strings, and not runtime dynamic content like we're handling here.

A better solution would be to add an attribute to the language in the API so that app updates are not required when adding future languages.