Remove buttonTextColor variable.
probably temporarily, need to deconstruct this to see how to put it back together.

The main problem is that every (most?) other elements that use a content:text element render a corresponding UILabel for it. However in the button case, we don't. The UIButton has a UILabel property already and we just have to set its attributes accordingly the XML spec. This difference, that is that a UILabel isn't created by our renderer, changes how the Button has to find and remember text attributes.