Don't remove objects on the fly and check for remaining objects.
The way swift Arrays work is that any array that is mutated while it's being iterated over is actually copied first. This means that we can safely iterate over the array and when it's done clear it.

Any new downloads added to the Array will be added to the copied Array and available for the next go around.