Move setting primary language earlier.
Settings this after the languages means setting the value only waits for one call to finish, not two. This deals with low-connectivity cases where the user might go select a language and then have their choice reverted once the network calls resolve.

Arguably... this could happen even sooner, like before the languages call is issued b/c we likely have most languages that a device would be set to already in the system and specified in the bundled languages.json , but that requires further analysis.